refactor: split game event factories into modules
