+++ src/utils/gameplay/gameEvents/movement.ts
@@ -1,0 +1,67 @@
+import {
+  Facing,
+  GameEventType,
+  GamePhase,
+  IGameEvent,
+  IHexCoordinate,
+  IMovementDeclaredPayload,
+  IMovementLockedPayload,
+  MovementType,
+} from '@/types/gameplay';
+
+import { createEventBase } from './base';
+
+export function createMovementDeclaredEvent(
+  gameId: string,
+  sequence: number,
+  turn: number,
+  unitId: string,
+  from: IHexCoordinate,
+  to: IHexCoordinate,
+  facing: Facing,
+  movementType: MovementType,
+  mpUsed: number,
+  heatGenerated: number,
+): IGameEvent {
+  const payload: IMovementDeclaredPayload = {
+    unitId,
+    from,
+    to,
+    facing,
+    movementType,
+    mpUsed,
+    heatGenerated,
+  };
+
+  return {
+    ...createEventBase(
+      gameId,
+      sequence,
+      GameEventType.MovementDeclared,
+      turn,
+      GamePhase.Movement,
+      unitId,
+    ),
+    payload,
+  };
+}
+
+export function createMovementLockedEvent(
+  gameId: string,
+  sequence: number,
+  turn: number,
+  unitId: string,
+): IGameEvent {
+  const payload: IMovementLockedPayload = { unitId };
+  return {
+    ...createEventBase(
+      gameId,
+      sequence,
+      GameEventType.MovementLocked,
+      turn,
+      GamePhase.Movement,
+      unitId,
+    ),
+    payload,
+  };
+}
